+++ src/modules/esl-toggleable/core.ts
@@ -1,2 +1,5 @@
+export type {ESLToggleableShape} from './core/esl-toggleable.shape';
+export type {ESLToggleableDispatcherShape} from './core/esl-toggleable-dispatcher.shape';
+
 export * from './core/esl-toggleable';
 export * from './core/esl-toggleable-dispatcher';